+++ packages/client/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 
+- Avoid race-conditions when destroying the client.
 - **CLI:**
   - Avoid throwing if there are no chains to generate.