+++ packages/client/CHANGELOG.md
@@ -5,13 +5,14 @@
 ### Fixed
 
 - Avoid race-conditions when destroying the client.
+- **CLI:**
+  - Avoid throwing if there are no chains to generate.
 
 ## 1.10.1 - 2025-05-04
 
 ### Fixed
 
 - Propagate the provided `asset` to the signer.
-
 - **CLI:**
   - Pin ES target version for descriptors.